[modify] 调整经纬度系数

+++ calc-engine/src/main/java/storm/domain/fence/service/impl/AbstractFenceQuery.java
@@ -63,7 +63,7 @@ public abstract class AbstractFenceQuery implements IFenceQueryService {
     /**
      * 经纬度坐标系中距离为1的两个坐标间近似距离93164米, 这里简化为1000000
      */
-    protected static final long COORDINATE_COEFFICIENT = ConfigUtils.getSysDefine().getFenceCoordinateCoefficient();
+    protected static final double COORDINATE_COEFFICIENT = ConfigUtils.getSysDefine().getFenceCoordinateCoefficient();
 
     private DataToRedis redis;
